clinker-core: convert grace_hash.rs to a module directory
Replace the 3,707-line pipeline/grace_hash.rs with a grace_hash/
directory whose mod.rs retains the high-level executor, the
execute_combine_grace_hash entry point, the partition state machine,
and the GraceHashConsumer memory-consumer wrapper, then re-exports the
build / probe / spill submodules. The unit-test harness moves to a
sibling tests.rs so each production submodule stays focused.

Pure relocation: no symbol renames and no public-API change. The
crate-facing surface (GraceHashExec, GraceHashExecutor,
execute_combine_grace_hash, GraceHashConsumer, GRACE_RECORD_BYTES_ESTIMATE)
and the Iterator impl on BuildChunkIter resolve through the same
pipeline::grace_hash paths external consumers already use. Tests are
relocated verbatim — no assertion changed.

Closes #130
